Injected interface refactor
