+++ src/main/java/com/fexl/circumnavigate/mixin/interfaceInjects/PlayerClientInjectorMixin.java
@@ -2,7 +2,11 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-package com.fexl.circumnavigate.mixin;
+/*
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+
+package com.fexl.circumnavigate.mixin.interfaceInjects;
 
 import com.fexl.circumnavigate.injected.ServerPlayerInjector;
 import net.minecraft.core.BlockPos;
@@ -29,10 +33,10 @@ public class PlayerClientInjectorMixin implements ServerPlayerInjector {
 	public ChunkPos getClientChunk() { return new ChunkPos((int) (clientX / 16), (int) (clientZ / 16)); }
 
 	@Override
-	public BlockPos getClientBlock() { return new BlockPos(Mth.floor(clientX), ((ServerPlayer)(Object)this).getBlockY(), Mth.floor(clientZ)); }
+	public BlockPos getClientBlock() { return new BlockPos(Mth.floor(clientX), ((ServerPlayer)(Object)this).blockPosition().getY(), Mth.floor(clientZ)); }
 
 	@Override
-	public Vec3 getClientPosition() { return new Vec3(clientX, ((ServerPlayer)(Object)this).getBlockY(), clientZ); }
+	public Vec3 getClientPosition() { return new Vec3(clientX, ((ServerPlayer)(Object)this).position().y, clientZ); }
 
 	@Override
 	public void setClientX(double clientX) {
